refactor: extract ClassicBrowserViewManager from ClassicBrowserService
- Created ClassicBrowserViewManager to handle WebContentsView lifecycle
- ClassicBrowserService now acts as a facade, delegating view operations
- Implemented event-driven communication between services via EventEmitter
- Maintained zero disruption to external API
- Single responsibility: ViewManager handles views, Service handles state

This is Phase 1 of the browser service refactoring to improve separation
of concerns and maintainability.
