+++ services/browser/ClassicBrowserViewManager.ts
@@ -1,0 +1,540 @@
+import { BrowserWindow, WebContentsView } from 'electron';
+import { EventEmitter } from 'events';
+import { BaseService } from '../base/BaseService';
+import { ClassicBrowserPayload, TabState } from '../../shared/types';
+import { CLASSIC_BROWSER_VIEW_FOCUSED } from '../../shared/ipcChannels';
+
+/**
+ * Dependencies for ClassicBrowserViewManager
+ */
+export interface ClassicBrowserViewManagerDeps {
+  mainWindow: BrowserWindow;
+  eventEmitter: EventEmitter;
+}
+
+/**
+ * ClassicBrowserViewManager
+ * 
+ * Responsible for the raw WebContentsView lifecycle:
+ * - Creating, destroying, and managing the physical state (bounds, visibility) of WebContentsView instances
+ * - Managing the views Map
+ * - Handling WebContentsView events and delegating them via EventEmitter
+ */
+export class ClassicBrowserViewManager extends BaseService<ClassicBrowserViewManagerDeps> {
+  private views: Map<string, WebContentsView> = new Map();
+
+  constructor(deps: ClassicBrowserViewManagerDeps) {
+    super('ClassicBrowserViewManager', deps);
+  }
+
+  /**
+   * Get a view by windowId
+   */
+  public getView(windowId: string): WebContentsView | undefined {
+    return this.views.get(windowId);
+  }
+
+  /**
+   * Get all window IDs that have active WebContentsViews
+   */
+  public getActiveViewWindowIds(): string[] {
+    return Array.from(this.views.keys());
+  }
+
+  /**
+   * Create a WebContentsView with the given state
+   */
+  public createViewWithState(windowId: string, bounds: Electron.Rectangle, browserState: ClassicBrowserPayload): void {
+    // Log Electron version
+    this.logDebug('Electron version:', process.versions.electron);
+
+    const securePrefs: Electron.WebPreferences = {
+      nodeIntegration: false,
+      contextIsolation: true,
+      sandbox: true,
+      preload: undefined, // Do not use the app's preload in the embedded browser
+      webSecurity: true,
+      // Enable plugins to support PDF viewer
+      plugins: true,
+    };
+
+    const view = new WebContentsView({ webPreferences: securePrefs });
+    this.views.set(windowId, view);
+
+    // Set initial background to transparent
+    view.setBackgroundColor('#00000000');
+    this.logDebug(`Set initial transparent background for window ${windowId}`);
+
+    // Apply border radius to the native view
+    // WebLayer uses 18px border radius, regular windows use 10px (12px - 2px border inset)
+    const borderRadius = windowId === '__WEBLAYER_SINGLETON__' ? 18 : 10;
+    (view as any).setBorderRadius(borderRadius);
+    this.logDebug(`✅ setBorderRadius called for windowId: ${windowId} with radius: ${borderRadius}px`);
+
+    if (!this.deps.mainWindow || this.deps.mainWindow.isDestroyed()) {
+      this.logError('Main window is not available to attach WebContentsView.');
+      this.views.delete(windowId); // Clean up
+      throw new Error('Main window not available.');
+    }
+
+    // Set bounds and add to window
+    view.setBounds(bounds);
+    this.deps.mainWindow.contentView.addChildView(view);
+
+    // Set up all WebContentsView event listeners
+    this.setupWebContentsListeners(windowId, view, browserState);
+
+    this.logDebug(`WebContentsView for windowId ${windowId} created and listeners attached.`);
+  }
+
+  /**
+   * Set up all event listeners for a WebContentsView
+   */
+  private setupWebContentsListeners(windowId: string, view: WebContentsView, browserState: ClassicBrowserPayload): void {
+    const wc = view.webContents;
+
+    // Inject CMD+Click handling script
+    const cmdClickInterceptorScript = `
+      (() => {
+        document.addEventListener('click', (event) => {
+          // Find the nearest ancestor 'a' tag
+          const link = event.target.closest('a');
+          if (link && (event.metaKey || event.ctrlKey)) {
+            // Prevent the default action (opening in the same view or a new OS window)
+            event.preventDefault();
+            
+            const targetUrl = link.href;
+            if (targetUrl) {
+              // Use a custom, non-existent protocol to send the URL to the main process.
+              // This will be caught by the 'will-navigate' event listener.
+              window.location.href = \`jeffers-ipc://cmd-click/\${encodeURIComponent(targetUrl)}\`;
+            }
+          }
+        }, true); // Use capture phase to catch the event early
+      })();
+    `;
+
+    wc.on('dom-ready', () => {
+      wc.executeJavaScript(cmdClickInterceptorScript).catch(err => {
+        this.logError(`windowId ${windowId}: Failed to inject CMD+click interceptor script:`, err);
+      });
+    });
+
+    // Loading events
+    wc.on('did-start-loading', () => {
+      this.logDebug(`windowId ${windowId}: did-start-loading`);
+      this.deps.eventEmitter.emit('view:did-start-loading', { windowId });
+    });
+
+    wc.on('did-stop-loading', () => {
+      this.logDebug(`windowId ${windowId}: did-stop-loading`);
+      this.deps.eventEmitter.emit('view:did-stop-loading', {
+        windowId,
+        url: wc.getURL(),
+        title: wc.getTitle(),
+        canGoBack: wc.navigationHistory.canGoBack(),
+        canGoForward: wc.navigationHistory.canGoForward(),
+      });
+    });
+
+    // Navigation events
+    wc.on('did-navigate', async (_event, url) => {
+      this.logDebug(`windowId ${windowId}: did-navigate to ${url}`);
+      this.deps.eventEmitter.emit('view:did-navigate', {
+        windowId,
+        url,
+        title: wc.getTitle(),
+        canGoBack: wc.navigationHistory.canGoBack(),
+        canGoForward: wc.navigationHistory.canGoForward(),
+      });
+    });
+
+    wc.on('did-navigate-in-page', async (_event, url, isMainFrame) => {
+      if (!isMainFrame) return;
+      this.logDebug(`windowId ${windowId}: did-navigate-in-page to ${url}`);
+      this.deps.eventEmitter.emit('view:did-navigate-in-page', {
+        windowId,
+        url,
+        title: wc.getTitle(),
+        canGoBack: wc.navigationHistory.canGoBack(),
+        canGoForward: wc.navigationHistory.canGoForward(),
+      });
+    });
+
+    // Title and favicon updates
+    wc.on('page-title-updated', (_event, title) => {
+      this.logDebug(`windowId ${windowId}: page-title-updated to ${title}`);
+      this.deps.eventEmitter.emit('view:page-title-updated', { windowId, title });
+    });
+
+    wc.on('page-favicon-updated', (_event, favicons) => {
+      this.logDebug(`windowId ${windowId}: page-favicon-updated with ${favicons.length} favicons`);
+      const faviconUrl = favicons.length > 0 ? favicons[0] : null;
+      this.deps.eventEmitter.emit('view:page-favicon-updated', { windowId, faviconUrl });
+    });
+
+    // Error handling
+    wc.on('did-fail-load', (_event, errorCode, errorDescription, validatedURL) => {
+      this.logDebug(`windowId ${windowId}: did-fail-load for ${validatedURL}. Code: ${errorCode}, Desc: ${errorDescription}`);
+      this.deps.eventEmitter.emit('view:did-fail-load', {
+        windowId,
+        errorCode,
+        errorDescription,
+        validatedURL,
+        currentUrl: wc.getURL(),
+        canGoBack: wc.navigationHistory.canGoBack(),
+        canGoForward: wc.navigationHistory.canGoForward(),
+      });
+    });
+
+    wc.on('render-process-gone', (_event, details) => {
+      this.logError(`windowId ${windowId}: render-process-gone. Reason: ${details.reason}`);
+      this.deps.eventEmitter.emit('view:render-process-gone', { windowId, details });
+    });
+
+    // Window open handling
+    wc.setWindowOpenHandler((details) => {
+      this.logDebug(`[setWindowOpenHandler] Intercepted window open request`, details);
+      this.deps.eventEmitter.emit('view:window-open-request', { windowId, details });
+      // The handler in ClassicBrowserService will decide what to do
+      return { action: 'deny' };
+    });
+
+    // Will-navigate handling (including CMD+click via custom protocol)
+    wc.on('will-navigate', (event, url) => {
+      this.deps.eventEmitter.emit('view:will-navigate', { windowId, event, url });
+    });
+
+    // Redirect events
+    wc.on('will-redirect', (event, url, isInPlace, isMainFrame) => {
+      this.logDebug(`windowId ${windowId}: will-redirect to ${url}, isInPlace: ${isInPlace}, isMainFrame: ${isMainFrame}`);
+    });
+
+    wc.on('did-redirect-navigation', (event, url, isInPlace, isMainFrame) => {
+      this.logDebug(`windowId ${windowId}: did-redirect-navigation to ${url}, isInPlace: ${isInPlace}, isMainFrame: ${isMainFrame}`);
+      this.deps.eventEmitter.emit('view:did-redirect-navigation', { windowId, url });
+    });
+
+    wc.on('did-start-navigation', (event, url, isInPlace, isMainFrame) => {
+      this.logDebug(`windowId ${windowId}: did-start-navigation to ${url}, isInPlace: ${isInPlace}, isMainFrame: ${isMainFrame}`);
+    });
+
+    // Iframe handling
+    wc.on('did-attach-webview', (event, webContents) => {
+      this.logDebug(`windowId ${windowId}: Attached webview, setting up handlers`);
+      webContents.setWindowOpenHandler((details) => {
+        this.logDebug(`windowId ${windowId}: Iframe intercepted new window request to ${details.url}`);
+        this.deps.eventEmitter.emit('view:iframe-window-open-request', { windowId, details });
+        return { action: 'deny' };
+      });
+    });
+
+    // Focus handling
+    wc.on('focus', () => {
+      this.logDebug(`windowId ${windowId}: WebContentsView received focus.`);
+      if (this.deps.mainWindow && !this.deps.mainWindow.isDestroyed()) {
+        this.deps.mainWindow.webContents.send(CLASSIC_BROWSER_VIEW_FOCUSED, { windowId });
+      }
+    });
+  }
+
+  /**
+   * Create a browser view - delegates to createViewWithState
+   */
+  public createBrowserView(windowId: string, bounds: Electron.Rectangle, payload: ClassicBrowserPayload): void {
+    this.createViewWithState(windowId, bounds, payload);
+  }
+
+  /**
+   * Destroy a specific browser view
+   */
+  public async destroyBrowserView(windowId: string): Promise<void> {
+    this.logDebug(`[DESTROY] Attempting to destroy WebContentsView for windowId: ${windowId}`);
+
+    const view = this.views.get(windowId);
+    if (!view) {
+      this.logWarn(`[DESTROY] No WebContentsView found for windowId ${windowId}. Nothing to destroy.`);
+      return;
+    }
+
+    // Remove from map immediately to prevent race conditions
+    this.views.delete(windowId);
+    this.logDebug(`[DESTROY] Found and removed view for ${windowId} from map. Proceeding with destruction.`);
+
+    // Detach from window if attached
+    if (this.deps.mainWindow && !this.deps.mainWindow.isDestroyed()) {
+      try {
+        if (this.deps.mainWindow.contentView && this.deps.mainWindow.contentView.children.includes(view)) {
+          this.deps.mainWindow.contentView.removeChildView(view);
+        }
+      } catch (error) {
+        this.logWarn(`[DESTROY] Error detaching view from window (might already be detached):`, error);
+      }
+    }
+
+    // Stop any media playback and cleanup before destroying
+    const wc = view.webContents;
+    if (wc && !wc.isDestroyed()) {
+      try {
+        // Mute audio immediately to stop sound
+        wc.setAudioMuted(true);
+
+        // Stop any pending loads
+        wc.stop();
+
+        // Only try to execute JavaScript if the page has loaded and not crashed
+        if (!wc.isLoading() && !wc.isCrashed()) {
+          try {
+            // Execute JavaScript to pause all media elements
+            await wc.executeJavaScript(`
+              (function() {
+                try {
+                  // Pause all video elements
+                  const videos = document.querySelectorAll('video');
+                  videos.forEach(video => {
+                    try {
+                      video.pause();
+                      video.currentTime = 0;
+                    } catch (e) {}
+                  });
+                  
+                  // Pause all audio elements
+                  const audios = document.querySelectorAll('audio');
+                  audios.forEach(audio => {
+                    try {
+                      audio.pause();
+                      audio.currentTime = 0;
+                    } catch (e) {}
+                  });
+                  
+                  // YouTube specific: stop via YouTube API if available
+                  if (typeof window !== 'undefined' && window.YT) {
+                    try {
+                      const players = document.querySelectorAll('.html5-video-player');
+                      players.forEach(player => {
+                        if (player.pauseVideo) player.pauseVideo();
+                      });
+                    } catch (e) {}
+                  }
+                } catch (e) {
+                  // Ignore errors, best effort cleanup
+                }
+                return true;
+              })();
+            `, true); // Add userGesture to be safe
+          } catch (scriptError) {
+            // Ignore script errors, continue with destruction
+            this.logDebug(`windowId ${windowId}: Script execution error during cleanup (ignored):`, scriptError);
+          }
+        }
+
+        // Small delay to ensure media cleanup takes effect
+        await new Promise(resolve => setTimeout(resolve, 50));
+
+        this.logDebug(`windowId ${windowId}: Stopped media playback and cleared page.`);
+      } catch (error) {
+        this.logWarn(`windowId ${windowId}: Error during media cleanup (ignored):`, error);
+      }
+    }
+
+    // Finally, destroy the webContents
+    if (view.webContents && !view.webContents.isDestroyed()) {
+      (view.webContents as any).destroy();
+    }
+
+    this.logDebug(`windowId ${windowId}: WebContentsView destruction process completed.`);
+  }
+
+  /**
+   * Destroy all browser views
+   */
+  public async destroyAllBrowserViews(): Promise<void> {
+    this.logDebug('Destroying all WebContentsViews.');
+    const destroyPromises = Array.from(this.views.keys()).map(windowId =>
+      this.destroyBrowserView(windowId)
+    );
+    await Promise.all(destroyPromises);
+  }
+
+  /**
+   * Set bounds for a view
+   */
+  public setBounds(windowId: string, bounds: Electron.Rectangle): void {
+    const view = this.views.get(windowId);
+    if (!view) {
+      // Don't warn for setBounds - this is expected during initialization
+      this.logDebug(`setBounds: No WebContentsView found for windowId ${windowId}. Skipping.`);
+      return;
+    }
+
+    // Validate bounds to prevent invalid values
+    if (!bounds || bounds.width <= 0 || bounds.height <= 0) {
+      this.logDebug(`setBounds: Invalid bounds for windowId ${windowId}. Skipping.`);
+      return;
+    }
+
+    this.logDebug(`windowId ${windowId}: Setting bounds to ${JSON.stringify(bounds)}`);
+    view.setBounds(bounds);
+  }
+
+  /**
+   * Set visibility for a view
+   */
+  public setVisibility(windowId: string, shouldBeDrawn: boolean, isFocused: boolean): void {
+    const view = this.views.get(windowId);
+    if (!view) {
+      // Don't warn for setVisibility - this might be called during cleanup
+      this.logDebug(`setVisibility: No WebContentsView found for windowId ${windowId}. Skipping.`);
+      return;
+    }
+
+    if (!this.deps.mainWindow || this.deps.mainWindow.isDestroyed()) {
+      this.logError('setVisibility: Main window is not available.');
+      return;
+    }
+
+    this.logDebug(`windowId ${windowId}: Setting visibility - shouldBeDrawn: ${shouldBeDrawn}, isFocused: ${isFocused}`);
+
+    const viewIsAttached = this.deps.mainWindow.contentView.children.includes(view);
+
+    if (shouldBeDrawn) {
+      // Make the view visible
+      (view as any).setVisible(true);
+
+      // The view ordering is now handled by the renderer through syncViewStackingOrder
+      // We only need to ensure the view is attached
+      if (!viewIsAttached) {
+        this.deps.mainWindow.contentView.addChildView(view);
+        this.logDebug(`windowId ${windowId}: Attached WebContentsView.`);
+      }
+    } else {
+      // Not to be drawn (e.g., minimized or window explicitly hidden)
+      (view as any).setVisible(false); // Make it not drawable
+      this.logDebug(`windowId ${windowId}: Set WebContentsView to not visible because shouldBeDrawn is false.`);
+      if (viewIsAttached) {
+        this.deps.mainWindow.contentView.removeChildView(view);
+        this.logDebug(`windowId ${windowId}: Removed WebContentsView from contentView because shouldBeDrawn is false.`);
+      }
+    }
+  }
+
+  /**
+   * Set the background color of the WebContentsView
+   */
+  public setBackgroundColor(windowId: string, color: string): void {
+    const view = this.views.get(windowId);
+    if (!view) {
+      this.logDebug(`setBackgroundColor: No WebContentsView found for windowId ${windowId}. Skipping.`);
+      return;
+    }
+
+    try {
+      view.setBackgroundColor(color);
+      this.logDebug(`Set background color for window ${windowId} to ${color}`);
+    } catch (error) {
+      this.logError(`Error setting background color for window ${windowId}:`, error);
+    }
+  }
+
+  /**
+   * Synchronize WebContentsView stacking order based on window z-indices
+   */
+  public syncViewStackingOrder(windowsInOrder: Array<{ id: string; isFrozen: boolean; isMinimized: boolean }>): void {
+    if (!this.deps.mainWindow || this.deps.mainWindow.isDestroyed()) {
+      this.logWarn('[syncViewStackingOrder] Main window is not available');
+      return;
+    }
+
+    this.logDebug(`[syncViewStackingOrder] Syncing view order for ${windowsInOrder.length} windows`);
+
+    // Build a list of views that should be visible
+    const viewsToShow: Array<{ windowId: string; view: WebContentsView }> = [];
+    const viewsToHide = new Set<string>();
+
+    // First pass: determine which views should be shown/hidden
+    for (const window of windowsInOrder) {
+      const view = this.views.get(window.id);
+      if (!view) continue;
+
+      const shouldBeVisible = !window.isFrozen && !window.isMinimized;
+
+      if (shouldBeVisible) {
+        viewsToShow.push({ windowId: window.id, view });
+      } else {
+        viewsToHide.add(window.id);
+      }
+    }
+
+    // Second pass: hide views that shouldn't be visible
+    for (const windowId of viewsToHide) {
+      const view = this.views.get(windowId);
+      if (view && this.deps.mainWindow.contentView.children.includes(view)) {
+        try {
+          this.deps.mainWindow.contentView.removeChildView(view);
+          this.logDebug(`[syncViewStackingOrder] Removed hidden view ${windowId}`);
+        } catch (error) {
+          this.logWarn(`[syncViewStackingOrder] Error removing view ${windowId}:`, error);
+        }
+      }
+    }
+
+    // Third pass: ensure visible views are in the correct order
+    const currentChildren = this.deps.mainWindow.contentView.children;
+    let needsReorder = false;
+
+    // Check if views are already in the correct order
+    if (currentChildren.length === viewsToShow.length) {
+      for (let i = 0; i < viewsToShow.length; i++) {
+        if (currentChildren[i] !== viewsToShow[i].view) {
+          needsReorder = true;
+          break;
+        }
+      }
+    } else {
+      needsReorder = true;
+    }
+
+    if (needsReorder) {
+      this.logDebug('[syncViewStackingOrder] Reordering views');
+
+      // Remove only the views that need to be reordered
+      for (const { view } of viewsToShow) {
+        if (this.deps.mainWindow.contentView.children.includes(view)) {
+          try {
+            this.deps.mainWindow.contentView.removeChildView(view);
+          } catch (error) {
+            this.logWarn('[syncViewStackingOrder] Error removing view for reorder:', error);
+          }
+        }
+      }
+
+      // Add views back in the correct order
+      for (const { windowId, view } of viewsToShow) {
+        try {
+          this.deps.mainWindow.contentView.addChildView(view);
+          this.logDebug(`[syncViewStackingOrder] Added view ${windowId} in correct position`);
+        } catch (error) {
+          this.logWarn(`[syncViewStackingOrder] Error adding view ${windowId}:`, error);
+        }
+      }
+    } else {
+      this.logDebug('[syncViewStackingOrder] Views already in correct order, no changes needed');
+    }
+
+    this.logDebug('[syncViewStackingOrder] View stacking order synchronized');
+  }
+
+  /**
+   * Clean up all resources when the service is destroyed
+   */
+  async cleanup(): Promise<void> {
+    try {
+      await this.destroyAllBrowserViews();
+    } catch (error) {
+      this.logError('Error destroying browser views during cleanup:', error);
+    }
+
+    this.views.clear();
+    this.logInfo('ClassicBrowserViewManager cleaned up');
+  }
+}
